Move providers into providers. Add bool, string, number, object getFeatureFlag variants for server side.
