+++ packages/datum-web/apps/operator/src/app/providers.tsx
@@ -5,7 +5,7 @@ import { createClient } from '@/lib/urql'
 import { useSession } from 'next-auth/react'
 import { ThemeProvider } from '@/providers/theme'
 import { ClientProvider } from '@/query/client-provider'
-import { FeatureFlagsProvider } from '@/lib/feature-flags/feature-flags-provider'
+import { FeatureFlagsProvider } from '@/providers/feature-flags-provider'
 interface ProvidersProps {
   children: any
 }
